Feat: 마이페이지 내가 쓴 피드, 댓글 adapter 수정 #74 id 가 long 형으로 바뀜에따라 피드 디테일로 보낼 id 형태 수정

+++ app/src/main/java/com/project/veganlife/mypage/ui/adapter/MypagePostedCommentAdapter.kt
@@ -55,7 +55,7 @@ class MypagePostedCommentAdapter :
 
                 clMypageLayout.setOnClickListener {
                     val bundle = Bundle()
-                    bundle.putInt("feedId", postedFeed.id)
+                    bundle.putLong("feedId", postedFeed.id)
                     itemView.findNavController().navigate(
                         R.id.action_mypagePostedCommentsFragment_to_communityDetailFeedFragment,
                         bundle
